add logic for publishing a test with question. move it to withCreateTestState.js along with validation logic
